feat: split game card into two components
